requirements-dev.txt
-r requirements.txt
pylint==2.3.1
invoke==1.2.0
mypy==0.730
pytest==5.0.1
pytest-asyncio==0.10.0
pytest-pythonpath==0.7.3
pytest-flakes==4.0.0
pytest-mock==1.10.4
PyGithub==1.44